перенёс отрисовку карты в MapService
