+++ src/components/offers-map/offers-map.jsx
@@ -1,4 +1,4 @@
-import leaflet from "leaflet";
+// import leaflet from "leaflet";
 import PropTypes from "prop-types";
 import React, {PureComponent} from "react";
 
@@ -13,37 +13,7 @@ export class OffersMap extends PureComponent {
 
   componentDidMount() {
     const {offers} = this.props;
-    const city = [52.38333, 4.9];
-    const icon = leaflet.icon({
-      iconUrl: `img/pin-map.svg`,
-      iconSize: [30, 30]
-    });
-
-    const zoom = 12;
-    try {
-      const map = leaflet.map(`map`, {
-        center: city,
-        zoom,
-        zoomControl: false,
-        marker: true
-      });
-      map.setView(city, zoom);
-
-      leaflet
-        .tileLayer(`https://{s}.basemaps.cartocdn.com/rastertiles/voyager/{z}/{x}/{y}{r}.png`, {
-          attribution: `&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors &copy; <a href="https://carto.com/attributions">CARTO</a>`
-        })
-        .addTo(map);
-
-      offers.map((offer) => {
-        const offerCords = offer.coordinates;
-        return leaflet
-          .marker(offerCords, {icon})
-          .addTo(map);
-      });
-    } catch (e) {
-      return;
-    }
+    this.props.service.render(offers);
   }
 }
 
@@ -56,5 +26,8 @@ OffersMap.propTypes = {
     isPremium: PropTypes.bool.isRequired,
     isFavorite: PropTypes.bool.isRequired,
     coordinates: PropTypes.arrayOf(PropTypes.number)
-  })).isRequired
+  })).isRequired,
+  service: PropTypes.shape({
+    render: PropTypes.func
+  }).isRequired
 };
